+++ packages/lumx-vue/src/components/badge/index.ts
@@ -1,0 +1,4 @@
+import Badge, { BadgeProps } from './Badge';
+
+export { Badge };
+export type { BadgeProps };